+++ ft_calloc.c
@@ -6,14 +6,13 @@
 /*   By: supersko <ndionis@student.42mulhouse.fr>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/27 14:44:37 by supersko          #+#    #+#             */
-/*   Updated: 2022/03/06 22:33:51 by supersko         ###   ########.fr       */
+/*   Updated: 2022/03/11 17:54:51 by supersko         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stdlib.h>
 #include "libft.h"
 
-void	*ft_calloc(unsigned int count, unsigned int size)
+void	*ft_calloc(size_t count, size_t size)
 {
 	void			*ptr;
 	unsigned int	ii;
