+++ ft_strjoin.c
@@ -6,7 +6,7 @@
 /*   By: supersko <ndionis@student.42mulhouse.fr>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/24 15:51:11 by supersko          #+#    #+#             */
-/*   Updated: 2022/03/10 16:17:27 by supersko         ###   ########.fr       */
+/*   Updated: 2022/03/11 18:22:54 by supersko         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strjoin(char const *s1, char const *s2)
 {
-	unsigned int	len_s1;
+	size_t			len_s1;
 	unsigned int	ii;
 	char			*ret;
